Реализовал активацию форм при загрузке карты на странице
