+++ js/main.js
@@ -1,34 +1,2 @@
 import {
-  activateMapFilters,
-  deactivateMapFilters
-} from './map-filters.js';
-
-import {
-  activateForm,
-  deactivateForm
-} from './form.js';
-
-import {
-} from './data.js';
-
-import {
-  map
 } from './map.js';
-
-
-const activatePage = () => {
-  activateMapFilters();
-  activateForm();
-};
-
-const deactivatePage = () => {
-  deactivateMapFilters();
-  deactivateForm();
-};
-
-deactivatePage();
-map.on('load', () => {
-  activatePage();
-});
-
-
